migrate Carto.Addin to Pro3

+++ src/ProSuite.Processing.AGP.Core/Domain/IProcessingContext.cs
@@ -39,13 +39,13 @@ namespace ProSuite.Processing.AGP.Core.Domain
 			ProcessingDataset dataset, Geometry extent = null, bool recycling = false);
 
 		/// <remarks>Honors <see cref="SelectionType"/></remarks>
-		int CountInputFeatures(ProcessingDataset dataset, Geometry extent = null);
+		long CountInputFeatures(ProcessingDataset dataset, Geometry extent = null);
 
 		/// <remarks>Ignores selection type, applies only given filter criteria</remarks>
 		IEnumerable<Feature> GetOtherFeatures(
 			ProcessingDataset dataset, Geometry extent = null, bool recycling = false);
 
 		/// <remarks>Ignores selection type, applies only given filter criteria</remarks>
-		int CountOtherFeatures(ProcessingDataset dataset, Geometry extent = null);
+		long CountOtherFeatures(ProcessingDataset dataset, Geometry extent = null);
 	}
 }
